Rename to `database_scan`
