+++ nano/node/CMakeLists.txt
@@ -49,8 +49,8 @@ add_library(
   bootstrap_ascending/throttle.cpp
   bootstrap_ascending/account_sets.hpp
   bootstrap_ascending/account_sets.cpp
-  bootstrap_ascending/iterators.hpp
-  bootstrap_ascending/iterators.cpp
+  bootstrap_ascending/database_scan.hpp
+  bootstrap_ascending/database_scan.cpp
   bootstrap_ascending/peer_scoring.hpp
   bootstrap_ascending/peer_scoring.cpp
   bootstrap_ascending/service.hpp
